CellTypeDataFetcher
This commit contains all the code required to fetch data using the raw GitHub link and feed it to the serialized fields in Unity so that the data can be freely accessed for the next stage of the CCF VR Gallery: Visualization.
This version only returns a result string, but the functionality exists, merely commented out for the sprint demo on 06/17/2022.